+++ src/imio/helpers/tests/test_email.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from collective.MockMailHost.MockMailHost import MockMailHost
 from imio.helpers.emailer import add_attachment
 from imio.helpers.emailer import create_html_email
 from imio.helpers.emailer import get_mail_host
@@ -53,8 +54,7 @@ class TestEmail(IntegrationTestCase):
         filepath = os.path.join(path, 'barcode.png')
         add_attachment(eml, 'barcode.png', filepath=filepath)
         mail_host = get_mail_host()
-        sec_send_orig = mail_host.secureSend
-        mail_host.secureSend = mail_host.send
+        MockMailHost.secureSend = MockMailHost.send
         mail_host.reset()
         send_email(eml, 'Email subject hé hé', 'noréply@from.org', 'dèst@to.org')
         self.assertIn('Subject: =?utf-8?q?Email_subject_h=C3=A9_h=C3=A9?=\n', mail_host.messages[0])
@@ -68,7 +68,6 @@ class TestEmail(IntegrationTestCase):
         self.assertTrue(send_email(eml, u'Email subject hé hé', u'noréply@from.org', u'dèst@to.org'))
         # not ok if in list
         self.assertRaises(UnicodeEncodeError, send_email, eml, u'Email subject', '<noreply@from.org>', ['dest@to.org', u'Stéphan Geulette <seg@to.org>'])
-        mail_host.secureSend = sec_send_orig
 
     def test_validate_email_address(self):
         self.assertTupleEqual(validate_email_address('name@domain.org'), (u'', u'name@domain.org'))